feat: Review를 생성할 때 Store의 NolgoatAverageGrade를 업데이트하는 기능 구현

+++ src/main/java/wad/seoul_nolgoat/domain/review/Review.java
@@ -42,7 +42,7 @@ public class Review extends BaseTimeEntity {
     }
 
     public Review(
-            int grade,
+            double grade,
             String content,
             String imageUrl,
             User user,
